+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+<a name="0.15.3"></a>
+## 0.15.3 (2019-04-28)
+
+
+### Features
+
+* **kinopoisk:** add `type` property to `getFilmInfo`
+* **kinopoisk:** add `kpTrailer` property to `getFilmInfo`
+
+
+
 <a name="0.15.2"></a>
 ## 0.15.2 (2019-04-27)